refactor: stories for UiHorizontalPaging
