--- src/components/organisms/UiHorizontalPaging/UiHorizontalPaging.mdx
@@ -1,28 +1,0 @@
-import { Canvas, Meta, Controls } from '@storybook/blocks';
-import Code from '../../../../docs/components/Code';
-import * as stories from './UiHorizontalPaging.stories.js';
-import scss from './UiHorizontalPaging.stories.scss?raw';
-
-<Meta of={stories}/>
-
-# UiHorizontalPaging
-<Canvas of={stories.Common}/>
-<Controls of={stories.Common}/>
-
-## Stories
-### With Default Slot
-<Canvas of={stories.WithDefaultSlot}/>
-
-### As Multilevel
-<Canvas of={stories.AsMultilevel}/>
-
-### As Mobile Menu
-This story support only mobile view. Please open this story in Canvas.
-
-<Canvas of={stories.AsMobileMenu}/>
-
-**Use following SCSS to get the same result**
-<Code
-  id="horizontal-paging-as-mobile-menu"
-  code={ scss }
-/>
