--- src/components/organisms/UiHorizontalPaging/UiHorizontalPaging.stories.scss
@@ -1,126 +1,0 @@
-@use "../../../styles/mixins";
-.language {
-  margin-block: calc(var(--space-12) * -1);
-  margin-inline: calc(var(--space-16) * -1);
-}
-
-.for-business {
-  &__products {
-    margin-block: var(--space-24) var(--space-32);
-  }
-
-  &__product {
-    @include mixins.override-logical(button, null, border-radius, 0);
-    @include mixins.override-logical(button, null, border-width, 0);
-
-    @include mixins.focus {
-      box-shadow: var(--focus-inner);
-    }
-
-    &-content {
-      display: flex;
-      flex-direction: column;
-      gap: var(--space-4);
-      flex: 1;
-      text-align: start;
-      margin-block-start: var(--space-4);
-    }
-
-    &-prefix-icon,
-    &-suffix-icon {
-      --icons-size: 2rem;
-    }
-
-    &-prefix-icon {
-      margin-inline: 0 var(--space-16);
-    }
-
-    &-suffix-icon {
-      margin-inline: var(--space-12) 0;
-    }
-  }
-}
-
-.medical-certification {
-  &__icon {
-    --icons-size: 3rem;
-    margin-block: var(--space-12) 0;
-  }
-
-  &__heading {
-    margin-block: var(--space-8);
-  }
-}
-
-.terms-of-service {
-  &__heading {
-    margin-block: 0 var(--space-32);
-  }
-
-  &__bullet-points {
-    --bullet-points-gap: var(--space-32);
-  }
-
-  &__sub-points {
-    --bullet-points-gap: var(--space-8);
-
-    margin-block-start: var(--space-12);
-  }
-}
-
-.interview-id {
-  &__heading {
-    margin-block: var(--space-4);
-  }
-}
-
-//<Code id="horizontal-paging-as-mobile-menu">
-.horizontal-paging-as-mobile-menu {
-  $element: horizontal-paging-as-mobile-menu;
-
-  --side-panel-content-padding-block: 0;
-  --side-panel-content-padding-inline: 0;
-
-  &__title {
-    display: flex;
-    gap: var(--space-12);
-  }
-
-  &__back {
-    align-self: flex-start;
-    margin-block-start: 3px;
-  }
-
-  &__menu {
-    @include mixins.use-logical($element + "-menu", padding, var(--space-12) var(--space-4));
-
-    flex: 0 0 100%;
-  }
-
-  &__language {
-    padding-inline: 0.625rem;
-    padding-block-end: var(--space-12);
-    margin-block-end: var(--space-4);
-    margin-inline: calc(var(--space-4) * -1);
-    border: solid var(--color-border-divider);
-    border-width: 0 0 1px 0;
-  }
-
-  &__log-out {
-    padding-inline: var(--space-16);
-    padding-block: var(--space-12);
-    margin-inline: calc(var(--space-8) * -1);
-    border: solid var(--color-border-divider);
-    border-width: 1px 0;
-
-    &::after {
-      border-block: 0;
-    }
-  }
-
-  &__footer {
-    padding-inline: var(--space-20);
-    padding-block: var(--space-24);
-  }
-}
-//</Code>
