+++ src/components/organisms/UiHorizontalPaging/stories/_internal/InterviewId.vue
@@ -1,0 +1,21 @@
+<template>
+  <UiHeading class="ui-heading--h4">
+    Interview ID:
+  </UiHeading>
+  <UiText class="interview-id__id">
+    8611d67x-063x-4390-x812-125288e0b070
+  </UiText>
+</template>
+
+<script setup lang="ts">
+import {
+  UiText,
+  UiHeading,
+} from '@infermedica/component-library';
+</script>
+
+<style lang="scss">
+.interview-id {
+ &__id {}
+}
+</style>
