+++ src/components/organisms/UiHorizontalPaging/stories/index.ts
@@ -1,0 +1,2 @@
+export { default as BasicStories } from './Basic.vue';
+export { default as BasicStoriesSource } from './Basic.vue?raw';
